+++ complex/src/lib.rs
@@ -25,9 +25,6 @@ extern crate rustc_serialize;
 #[cfg(feature = "serde")]
 extern crate serde;
 
-#[cfg(feature = "quickcheck")]
-pub mod quickcheck_impls;
-
 use std::fmt;
 #[cfg(test)]
 use std::hash;
@@ -743,6 +740,64 @@ impl<T> fmt::Binary for Complex<T> where
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write_complex!(f, "b", "0b", self.re, self.im, T)
+    }
+}
+
+#[cfg(feature = "quickcheck")]
+impl<T: quickcheck::Arbitrary + Num> quickcheck::Arbitrary for Complex<T> {
+    fn arbitrary<G: quickcheck::Gen>(g: &mut G) -> Self {
+        Complex::new(T::arbitrary(g), T::arbitrary(g))
+    }
+
+    fn shrink(&self) -> Box<Iterator<Item = Self>> {
+        use quickcheck::Arbitrary;
+
+        struct Shrinker<T: Arbitrary + Num> {
+            mode: bool,
+            re: T,
+            im: T,
+            re_iter: Option<Box<Iterator<Item = T>>>,
+            im_iter: Option<Box<Iterator<Item = T>>>,
+        }
+        
+        impl<T: Arbitrary + Num> Shrinker<T> {
+            fn new(re: T, im: T) -> Self {
+                Shrinker {
+                    mode: true,
+                    re: re.clone(),
+                    im: im,
+                    re_iter: Some(re.shrink()),
+                    im_iter: None,
+                }
+            }
+        }
+        
+        impl<T: Arbitrary + Num> Iterator for Shrinker<T> {
+            type Item = Complex<T>;
+        
+            fn next(&mut self) -> Option<Complex<T>> {
+                match self.mode {
+                    true => if let Some(re) = self.re_iter.as_mut().unwrap().next() {
+                        Some(Complex::new(re, self.im.clone()))
+                    } else {
+                        self.mode = false;
+                        self.re_iter = None;
+                        self.im_iter = Some(self.im.clone().shrink());
+                        self.next()
+                    },
+                    false => if let Some(im) = self.im_iter.as_mut().unwrap().next() {
+                        Some(Complex::new(self.re.clone(), im))
+                    } else if let Some(re) = self.re_iter.as_mut().unwrap().next() {
+                        self.re = re;
+                        self.next()
+                    } else {
+                        None
+                    }
+                }
+            }
+        }
+
+        Box::new(Shrinker::new(self.re.clone(), self.im.clone()))
     }
 }
